refactor: px to rem, seperate styles and types
