+++ src/components/ConfirmationModal/confirmationModal.style.tsx
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+const ModalText = styled.p`
+  font-size: 1rem;
+  margin: 0;
+  text-align: center;
+  padding: 1rem;
+`;
+
+const ModalActions = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0.625rem;
+  margin-bottom: 1rem;
+`;
+
+export { ModalActions, ModalText };
