+++ src/components/Form/form.style.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const ButtonContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+`;
+export { ButtonContainer };
