+++ src/components/Spinner/Spinner.tsx
@@ -1,0 +1,10 @@
+import { Spinner, SpinnerWrapper } from "./spinner.style";
+
+// Component definition
+export default function SpinnerComponent() {
+  return (
+    <SpinnerWrapper>
+      <Spinner />
+    </SpinnerWrapper>
+  );
+}
